# docs/swagger.yaml
basePath: /api/v1.
definitions:
  APIError:
    properties:
      data:
        $ref: '#/definitions/errorStruct'
      ok:
        type: boolean
    type: object
  APIResponse-Group:
    properties:
      data:
        $ref: '#/definitions/Group'
      ok:
        type: boolean
    type: object
  APIResponse-IDResponse:
    properties:
      data:
        $ref: '#/definitions/IDResponse'
      ok:
        type: boolean
    type: object
  APIResponse-MessageResponse:
    properties:
      data:
        $ref: '#/definitions/MessageResponse'
      ok:
        type: boolean
    type: object
...
  APIResponse-array_Group:
    properties:
      data:
        items:
          $ref: '#/definitions/Group'
        type: array
      ok:
        type: boolean
    type: object
...
  APIResponse-array_Submission:
    properties:
      data:
        items:
          $ref: '#/definitions/Submission'
        type: array
      ok:
        type: boolean
    type: object
  APIResponse-array_Task:
    properties:
      data:
        items:
          $ref: '#/definitions/Task'
        type: array
      ok:
        type: boolean
    type: object
  APIResponse-array_User:
    properties:
      data:
        items:
          $ref: '#/definitions/User'
        type: array
      ok:
        type: boolean
    type: object
...
  CreateGroup:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
  EditGroup:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  Group:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  IDResponse:
    properties:
      id:
        type: integer
    type: object
  LanguageConfig:
    properties:
      fileExtension:
        type: string
      id:
        type: integer
      language:
        type: string
      version:
        type: string
    type: object
  MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  Submission:
    properties:
      checkedAt:
        type: string
      id:
        type: integer
      language:
        $ref: '#/definitions/LanguageConfig'
      languageId:
        type: integer
      order:
        type: integer
      result:
        $ref: '#/definitions/SubmissionResult'
      status:
        type: string
      submittedAt:
        type: string
      task:
        $ref: '#/definitions/Task'
      taskId:
        type: integer
      user:
        $ref: '#/definitions/User'
      userId:
        type: integer
    type: object
  SubmissionResult:
    properties:
      code:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      message:
        type: string
      submissionId:
        type: integer
      testResults:
        items:
          $ref: '#/definitions/TestResult'
        type: array
    type: object
  Task:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  TestResult:
    properties:
      id:
        type: integer
      inputOutputId:
        type: integer
      passed:
        type: boolean
      submissionResultId:
        type: integer
    type: object
  User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/types.UserRole'
      surname:
        type: string
      username:
        type: string
    type: object
...
  UserIDs:
    properties:
      userIDs:
        items:
          type: integer
        type: array
    type: object
...
  errorStruct:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  types.UserRole:
    enum:
    - student
    - teacher
    - admin
    type: string
    x-enum-varnames:
    - UserRoleStudent
    - UserRoleTeacher
    - UserRoleAdmin
info:
  contact: {}
  title: Mini-Maxit API
  version: 1.0.0
paths:
...
  /group/:
    get:
      description: Get all groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-array_Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get all groups
      tags:
      - group
    post:
      consumes:
      - application/json
      description: Create a group
      parameters:
      - description: Create Group
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Create a group
      tags:
      - group
  /group/{id}:
    get:
      description: Get a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get a group
      tags:
      - group
    put:
      consumes:
      - application/json
      description: Edit a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Edit Group
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/EditGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Edit a group
      tags:
      - group
  /group/{id}/users:
    delete:
      consumes:
      - application/json
      description: Delete users from a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: User IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserIDs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Delete users from a group
      tags:
      - group
    get:
      description: Get users in a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-array_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get users in a group
      tags:
      - group
    post:
      consumes:
      - application/json
      description: Add users to a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: User IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserIDs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Add users to a group
      tags:
      - group
...
  /submission/group/{id}:
    get:
      description: If the user is a student, it fails with 403 Forbidden.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit the number of returned submissions
        in: query
        name: limit
        type: integer
      - description: Offset the returned submissions
        in: query
        name: offset
        type: integer
      - description: Session Token
        in: header
        name: Session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-array_Submission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get all submissions for a group
      tags:
      - submission
...
  /task/{id}/assign/groups:
    post:
      description: Assigns a task to groups by task ID and group IDs
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group IDs
        in: body
        name: groupIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Assign a task to groups
      tags:
      - task
...
  /task/{id}/unassign/groups:
    delete:
      description: Unassigns a task from groups by task ID and group IDs
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group IDs
        in: body
        name: groupIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Unassign a task from groups
      tags:
      - task
...
  /task/group/{id}:
    get:
      description: Returns all tasks for a group by ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-array_Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get all tasks for a group
      tags:
      - task
...
swagger: "2.0"